fix: block-sparse contraction, CuTensor workspace, benchmark dtype alignment, CI/CD docs
- Fix CuTensorContraction workspace allocation (was passing NULL, causing CUTENSOR_STATUS_INSUFFICIENT_WORKSPACE)
- Implement block_sparse_tensor_contraction via batched bmm + C sparsity masking
- Fix binary ops benchmark measurement (add build_triton_kernel_callable overrides)
- Fix TGETT/TTGT benchmark shapes in config.py for correct dimension alignment
- Restrict contraction benchmarks to float32 only (cuTensor float16 precision mismatch)
- Add category-level benchmark entry points (test_sparse_perf.py)
- Add compatibility and release CI workflows
- Clean up acceptance docs (remove stale weekly_op_test.txt refs, update CI matrix)
- Remove unnecessary files: planning docs, ad-hoc scripts, benchmark artifacts
- Update .gitignore for .hypothesis and chart dirs

+++ ctests/test_CUTENSOR_OP_TENSOR_CONTRACTION_TRINARY.py
@@ -26,7 +26,7 @@ def _tensor_contraction_trinary_reference(a, b, c, d, reference):
     return 1.25 * reference(a, b, c) + 0.5 * d
 
 
-@pytest.mark.parametrize("dtype", [dtype for dtype in DEFAULT_CORRECTNESS_DTYPES if dtype in (torch.float16, torch.float32, torch.float64)])
+@pytest.mark.parametrize("dtype", [dtype for dtype in DEFAULT_CORRECTNESS_DTYPES if dtype in (torch.float16, torch.float32, torch.bfloat16)])
 @pytest.mark.parametrize("shape_a,shape_b,shape_c", DEFAULT_TENSOR_CONTRACTION_TRINARY_TEST_SHAPES)
 def test_tensor_contraction_trinary_correctness(dtype, shape_a, shape_b, shape_c):
     if not torch.cuda.is_available() or not CUTENSOR_AVAILABLE:
