+++ ctests/test_CUTENSOR_OP_TTGT.py
@@ -21,7 +21,7 @@ def _ttgt_reference(a, b, c, reference):
     return 1.25 * reference(a, b) + 0.5 * c
 
 
-@pytest.mark.parametrize("dtype", [dtype for dtype in DEFAULT_CORRECTNESS_DTYPES if dtype in (torch.float16, torch.float32, torch.float64)])
+@pytest.mark.parametrize("dtype", [dtype for dtype in DEFAULT_CORRECTNESS_DTYPES if dtype in (torch.float16, torch.float32, torch.bfloat16)])
 @pytest.mark.parametrize("shape_a,shape_b", DEFAULT_TTGT_TEST_SHAPES)
 def test_ttgt_correctness(dtype, shape_a, shape_b):
     if not torch.cuda.is_available() or not CUTENSOR_AVAILABLE:
